refactor: extract ScreenshotSelectionGeometry.swift
Extracts the public ScreenshotSelectionGeometry struct (21 lines) from
ScreenshotSelectionService.swift into its own file.

Byte-identical move; SHA-256 8555ffbdd8df96370ae016fb34e4a8fb623221a1bf633c2076d784242b13cb4d
verified against origin/main:ScreenshotSelectionService.swift lines 4-24.

Reduces ScreenshotSelectionService.swift from 396 to 373 lines.

Tests: 667 (unchanged).
